feat: Featured app cards and better spacing
