+++ packages/app_center/lib/src/games/games_page.dart
@@ -4,6 +4,7 @@ import 'package:app_center/l10n.dart';
 import 'package:app_center/layout.dart';
 import 'package:app_center/snapd.dart';
 import 'package:app_center/store.dart';
+import 'package:app_center/widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
@@ -31,8 +32,10 @@ class GamesPage extends ConsumerWidget {
           Text("Here we'll have featured games!"),
           SizedBox(height: kPagePadding),
         ]),
+        const CategorySnapList(category: SnapCategoryEnum.games),
         SliverList.list(
           children: [
+            const SizedBox(height: 56),
             Row(
               children: [
                 const Expanded(
@@ -83,6 +86,9 @@ class GamesPage extends ConsumerWidget {
                       fontSize: _CategoryBannerProperties.fontSize,
                     ))
               ],
+            ),
+            const SizedBox(
+              height: kPagePadding,
             )
           ],
         ),
