+++ packages/app_center/lib/widgets.dart
@@ -2,6 +2,7 @@ export 'src/widgets/app_card.dart';
 export 'src/widgets/app_icon.dart';
 export 'src/widgets/app_info_bar.dart';
 export 'src/widgets/app_title.dart';
+export 'src/widgets/category_snap_list.dart';
 export 'src/widgets/clickable.dart';
 export 'src/widgets/dialogs.dart';
 export 'src/widgets/screenshot_gallery.dart';
